refactor: 사용하지 않는 signaling stomp 코드를 제거합니다.
