+++ src/frontend/src/layouts/main-layout/index.tsx
@@ -12,7 +12,7 @@ import useGetSelfUser from '@/hooks/queries/user/useGetSelfUser'
 import { useChattingStomp } from '@/hooks/use-chatting-stomp'
 import { useMediaSettingsStore } from '@/stores/use-media-setting.store'
 import { useServerUnreadStore } from '@/stores/use-server-unread-store'
-import { useSignalingSocket } from '@/stores/use-signaling-socket'
+import { useSignalingSocket } from '@/stores/use-signaling-store'
 
 import ProfileCard from './components/profile-card'
 import ProfileStatusButton from './components/profile-status-button'
